Reorganize traits
