+++ lang/syntax/src/ast/decls.rs
@@ -11,6 +11,7 @@ use super::exp::*;
 use super::ident::*;
 use super::lookup_table::{DeclKind, LookupTable};
 use super::subst::{Substitutable, Substitution};
+use super::traits::HasSpan;
 
 #[derive(Debug, Clone)]
 pub struct DocComment {
