+++ lang/syntax/src/ast/traits/mod.rs
@@ -1,0 +1,7 @@
+mod has_span;
+mod has_type;
+mod occurs;
+
+pub use has_span::*;
+pub use has_type::*;
+pub use occurs::*;
