+++ lang/syntax/src/ast/traits/occurs.rs
@@ -1,13 +1,8 @@
 use std::rc::Rc;
 
+use crate::ast::exp::Exp;
 use crate::common::{Idx, Leveled, Lvl};
 use crate::ctx::LevelCtx;
-
-use super::exp::Exp;
-
-// Occurs
-//
-//
 
 pub trait Occurs {
     fn occurs(&self, ctx: &mut LevelCtx, lvl: Lvl) -> bool;
@@ -34,12 +29,4 @@ impl<T: Occurs> Occurs for Option<T> {
     fn occurs(&self, ctx: &mut LevelCtx, lvl: Lvl) -> bool {
         self.as_ref().map(|inner| inner.occurs(ctx, lvl)).unwrap_or_default()
     }
-}
-
-// HasTypeInfo
-//
-//
-
-pub trait HasTypeInfo {
-    fn typ(&self) -> Option<Rc<Exp>>;
 }
